+++ android/src/main/java/io/adtrace/sdk/flutter/AdTraceUtils.java
@@ -1,6 +1,4 @@
-//
-//  Created by Aref Hosseini on 7th October 2019.
-//
+
 
 package io.adtrace.sdk.flutter;