+++ example/android/app/src/main/java/io/adtrace/sdk/flutter/adtrace_sdk_flutter_example/MainActivity.java
@@ -1,0 +1,6 @@
+package io.adtrace.sdk.flutter.adtrace_sdk_flutter_example;
+
+import io.flutter.embedding.android.FlutterActivity;
+
+public class MainActivity extends FlutterActivity {
+}