+++ ios/Classes/AdTraceSdk.h
@@ -1,8 +1,4 @@
-//
-//  Created by Aref Hosseini on 7th October 2019.
-//
-
 #import <Flutter/Flutter.h>
 
-@interface AdTraceSdk : NSObject<FlutterPlugin>
+@interface AdtraceSdk : NSObject<FlutterPlugin>
 @end