+++ lib/adtrace_event_failure.dart
@@ -1,30 +1,43 @@
-//
-//  Created by Aref Hosseini on 7th October 2019.
-//
+/**
+ *  AdTrace SDK for Flutter
+ *  Developed by Nasser Amini on Mar 2021
+ *  for more information visit https://adtrace.io
+ */
 
 class AdTraceEventFailure {
-  String message;
-  String timestamp;
-  String adid;
-  String eventToken;
-  String callbackId;
-  String jsonResponse;
-  bool willRetry;
+  final String? message;
+  final String? timestamp;
+  final String? adid;
+  final String? eventToken;
+  final String? callbackId;
+  final String? jsonResponse;
+  final bool? willRetry;
 
-  static AdTraceEventFailure fromMap(dynamic map) {
-    AdTraceEventFailure eventFailure = new AdTraceEventFailure();
+  AdTraceEventFailure({
+    required this.message,
+    required this.timestamp,
+    required this.adid,
+    required this.eventToken,
+    required this.callbackId,
+    required this.jsonResponse,
+    required this.willRetry,
+  });
+
+  factory AdTraceEventFailure.fromMap(dynamic map) {
     try {
-      eventFailure.message = map['message'];
-      eventFailure.timestamp = map['timestamp'];
-      eventFailure.adid = map['adid'];
-      eventFailure.eventToken = map['eventToken'];
-      eventFailure.callbackId = map['callbackId'];
-      eventFailure.jsonResponse = map['jsonResponse'];
-      bool willRetry = map['willRetry'].toString().toLowerCase() == 'true';
-      eventFailure.willRetry = willRetry;
+      return AdTraceEventFailure(
+        message: map['message'],
+        timestamp: map['timestamp'],
+        adid: map['adid'],
+        eventToken: map['eventToken'],
+        callbackId: map['callbackId'],
+        jsonResponse: map['jsonResponse'],
+        willRetry: map['willRetry']?.toString().toLowerCase() == 'true',
+      );
     } catch (e) {
-      print('[AdTraceFlutter]: Failed to create AdTraceEventFailure object from given map object. Details: ' + e.toString());
+      throw Exception(
+          '[AdTraceFlutter]: Failed to create AdTraceEventFailure object from given map object. Details: ' +
+              e.toString());
     }
-    return eventFailure;
   }
 }