+++ lib/adtrace_session_failure.dart
@@ -1,26 +1,37 @@
-//
-//  Created by Aref Hosseini on 7th October 2019.
-//
+/**
+ *  AdTrace SDK for Flutter
+ *  Developed by Nasser Amini on Mar 2021
+ *  for more information visit https://adtrace.io
+ */
 
 class AdTraceSessionFailure {
-  String message;
-  String timestamp;
-  String adid;
-  String jsonResponse;
-  bool willRetry;
+  final String? message;
+  final String? timestamp;
+  final String? adid;
+  final String? jsonResponse;
+  final bool? willRetry;
 
-  static AdTraceSessionFailure fromMap(dynamic map) {
-    AdTraceSessionFailure sessionFailure = new AdTraceSessionFailure();
+  AdTraceSessionFailure({
+    required this.message,
+    required this.timestamp,
+    required this.adid,
+    required this.jsonResponse,
+    required this.willRetry,
+  });
+
+  factory AdTraceSessionFailure.fromMap(dynamic map) {
     try {
-      sessionFailure.message = map['message'];
-      sessionFailure.timestamp = map['timestamp'];
-      sessionFailure.adid = map['adid'];
-      sessionFailure.jsonResponse = map['jsonResponse'];
-      bool willRetry = map['willRetry'].toString().toLowerCase() == 'true';
-      sessionFailure.willRetry = willRetry;
+      return AdTraceSessionFailure(
+        message: map['message'],
+        timestamp: map['timestamp'],
+        adid: map['adid'],
+        jsonResponse: map['jsonResponse'],
+        willRetry: map['willRetry']?.toString().toLowerCase() == 'true',
+      );
     } catch (e) {
-      print('[AdTraceFlutter]: Failed to create AdTraceSessionFailure object from given map object. Details: ' + e.toString());
+      throw Exception(
+          '[AdTraceFlutter]: Failed to create AdTraceSessionFailure object from given map object. Details: ' +
+              e.toString());
     }
-    return sessionFailure;
   }
 }