+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+## 1.1.0+3
+* Readme updated
+
+## 1.1.0
+* AdTrace iOS SDK 2.0.7 integration
+
+## 1.0.3
+* bug fixes
+* update readme
+
+## 1.0.2
+
+* AdTrace android sdk version 2.0.2 integration
+* AdTrace iOS sdk version 2.0.5 DEV mode (**TEST**) integration (**STABLE will be published ASAP**)
+* `null-safety`
+* Major changes and bug fixes
+
 ## 0.1.3
 
 * Migrate to AndroidX.